feat(dashboard): move Group Sessions from left nav into a dashboard tab
Both the student and tutor dashboards gain a "Group Sessions" tab in their
SessionCalendarPanel, and the "Group Sessions" item is removed from each
left nav.

- Extract each group-sessions page body into a reusable panel component
  (student-/tutor-group-sessions-panel) with an `embedded` prop that drops
  the standalone hero (and, for the tutor, flattens the inner card) when
  rendered inside a dashboard tab.
- The standalone /group-sessions routes stay live (they now render the same
  panel) so existing deep links keep working.
- Student dashboard: the group-seat "Complete payment" button now switches
  to the Group Sessions tab instead of navigating to the old page.
- Remove the now-unused `Users` nav icon import from both layouts.

typecheck + build + 709 unit tests pass.
