+++ tutorme-app/src/app/[locale]/student/layout.tsx
@@ -17,7 +17,6 @@ import {
   Bell,
   ArrowLeft,
   CalendarDays,
-  Users,
   User,
   ChevronRight,
   ChevronLeft,
@@ -41,12 +40,6 @@ const navItems: NavItem[] = [
     label: 'Dashboard',
     icon: LayoutDashboard,
     iconColor: 'text-[#2563EB]',
-  },
-  {
-    href: '/student/group-sessions',
-    label: 'Group Sessions',
-    icon: Users,
-    iconColor: 'text-[#0891B2]',
   },
   {
     href: '/student/tutors',
