+++ tutorme-app/src/app/[locale]/tutor/dashboard/page.tsx
@@ -60,6 +60,7 @@ import {
 } from './components'
 import { DEFAULT_TIMEZONE, type CalendarView } from './components/InteractiveCalendar'
 import { SessionCalendarPanel } from '@/components/session-calendar-panel'
+import { TutorGroupSessionsPanel } from '@/components/group-session/tutor-group-sessions-panel'
 import { ModernHeroSection } from './components/ModernHeroSection'
 import { TutorPendingRescheduleBanner } from './components/TutorPendingRescheduleBanner'
 import { CountryFlag } from '@/components/country-flag'
@@ -766,6 +767,7 @@ function TutorDashboardContent() {
               { value: 'calendar', label: 'Calendar' },
               { value: 'availability', label: 'My Availability' },
               { value: 'oneOnOne', label: '1-on-1 Requests' },
+              { value: 'groupSessions', label: 'Group Sessions' },
             ]}
             showCalendarControls={activeTab === 'calendar' || activeTab === 'availability'}
             calendarView={calendarView}
@@ -1114,6 +1116,13 @@ function TutorDashboardContent() {
                   )}
                 </div>
               </div>
+            </TabsContent>
+
+            <TabsContent
+              value="groupSessions"
+              className="mt-0 flex min-h-0 flex-1 flex-col overflow-hidden"
+            >
+              <TutorGroupSessionsPanel embedded />
             </TabsContent>
           </SessionCalendarPanel>
         </div>
