+++ tutorme-app/src/app/[locale]/tutor/layout.tsx
@@ -23,7 +23,6 @@ import {
   ClipboardCheck,
   LogOut,
   User,
-  Users,
   FileText,
   ChevronRight,
   ChevronLeft,
@@ -61,12 +60,6 @@ const navItems: NavItem[] = [
     label: 'Grading',
     icon: ClipboardCheck,
     iconColor: 'text-[#16A34A]',
-  },
-  {
-    href: '/tutor/group-sessions',
-    label: 'Group Sessions',
-    icon: Users,
-    iconColor: 'text-[#0891B2]',
   },
   { href: '/tutor/reports', label: 'Analytics', icon: BarChart3, iconColor: 'text-[#F59E0B]' },
   { href: '/tutor/support', label: 'Support', icon: HelpCircle, iconColor: 'text-[#8B5CF6]' },
